增加对friendinfo的增、删、查操作（没有改！！！

+++ authentication/connecter/SQLConn.py
@@ -1,5 +1,6 @@
 from time import strftime, localtime
 
+from authentication.constantName import *
 from authentication.connecter.ServerConn import ServerConnect
 
 
@@ -10,30 +11,46 @@ def get_localtime():
 class ConnectSQL():
     def __init__(self, server_address):
         self.cur = ServerConnect(server_address)
-        self.property_name = ['id', 'name', 'mail', 'password', 'anime', 'profile']
+        self.property_name_userinfo = ['id', 'name', 'mail', 'password', 'anime', 'profile']
+        self.property_name_friendinfo = ['id', 'friend']
 
     def insert(self, table_name, user_data):
         """
         向数据库插入数据，不允许留空
         :param : table_name '需要操作的表名'
-        :param : user_data['name','mail','password']
+        :param : user_data['name','mail','password']--userinfo
+                 user_data['id', 'friend']--friendinfo
         :return: 插入失败则返回False
         """
-        try:
-            sql = f'select * from {table_name};'
-            send_data = {'sql': sql, 'args': None}
-            self.cur.send_sql('search', send_data)
-            current_id = str(self.cur.get_count() + 1)
-            print(current_id)
-            # 将用户提交的昵称、邮箱地址、密码提交,用户的ID为当前表行数+1
-            sql = f'insert into {table_name}(id,name, mail, password) values({current_id}, %s, %s, %s);'
-            data = {'sql': sql, 'args': user_data}
-            self.cur.send_sql('insert', data)
-            print(f'提交成功，更新{self.cur.get_count()}条数据')
-        except Exception as e:
-            # 发生错误
-            print(f"提交新用户数据发生错误:{e}")
-            return False
+        if table_name == TABLE_NAME_USERINFO:
+            try:
+                sql = f'select * from {table_name};'
+                send_data = {'sql': sql, 'args': None}
+                self.cur.send_sql('search', send_data)
+                current_id = str(self.cur.get_count() + 1)
+                print(current_id)
+                # 将用户提交的昵称、邮箱地址、密码提交,用户的ID为当前表行数+1
+                sql = f'insert into {table_name}(id,name, mail, password) values({current_id}, %s, %s, %s);'
+                data = {'sql': sql, 'args': user_data}
+                self.cur.send_sql('insert', data)
+                print(f'提交成功，更新{self.cur.get_count()}条数据')
+            except Exception as e:
+                # 发生错误
+                print(f"提交数据到{table_name}发生错误:{e}")
+                return False
+
+        if table_name == TABLE_NAME_FRIENDINFO:
+            try:
+                #
+                sql = f'insert into {table_name}(id,friend) values(%s, %s);'
+                data = {'sql': sql, 'args': user_data}
+                self.cur.send_sql('insert', data)
+                print(f'提交成功，更新{self.cur.get_count()}条数据')
+            except Exception as e:
+                # 发生错误
+                print(f"提交数据到{table_name}发生错误:{e}")
+                return False
+
 
     def search(self, table_name, data):
         """
@@ -41,8 +58,14 @@ class ConnectSQL():
         :param table_name '需要操作的表名'
         :param data['属性名', '该属性需查找的内容']
         :return results:一个n维数组[n][id, name, mail, password] 未查到则返回 None,查询失败而返回False
+                ans：一个列表[]，该id所有的好友id 未查到则返回 None,查询失败而返回False
         """
-        for i in self.property_name:
+        property_name = [' ']
+        if table_name == TABLE_NAME_USERINFO:
+            property_name = self.property_name_userinfo
+        if table_name == TABLE_NAME_FRIENDINFO:
+            property_name = self.property_name_friendinfo
+        for i in property_name:
             if i == data[0]:
                 try:
                     sql = f'select * from {table_name} where {data[0]} = %s;'
@@ -51,7 +74,13 @@ class ConnectSQL():
                     if self.cur.get_count():
                         result = self.cur.get_result()
                         print(f'查找成功,共查找到{self.cur.get_count()}条数据')
-                        return result
+                        if table_name == TABLE_NAME_USERINFO:
+                            return result
+                        if table_name == TABLE_NAME_FRIENDINFO:
+                            ans = []
+                            for i in result:
+                                ans.append(i[1])
+                            return ans
                     else:
                         print(f'没有查找到{data[0]}为{data[1]}的用户')
                         return None
@@ -67,26 +96,31 @@ class ConnectSQL():
         """
         删除数据库中的指定数据,data数据留空则删除最后一条
         :param table_name '需要操作的表名'
-        :param data['属性名', '该属性需删除的内容']
+        :param data['属性名', '该属性需删除的内容']--userinfo
+               data[id， friend]--friendinfo
         :return results:删除的数量  发生错误则返回NFalse
         """
-        try:
-            # if data:
-            #     # 无参时，删除数据库中的最后一条
-            #     sql = f'delete from {table_name} where id like (select top 1 id from {table_name} order by id desc)'
-            #     send_data = {'sql': sql, 'args': None}
-            #     self.cur.send_sql('delete', send_data)
-            #     return self.cur.get_count()
+        if table_name == TABLE_NAME_USERINFO:
+            try:
+                sql = f'delete from {table_name} where {data[0]} = %s;'
+                send_data = {'sql': sql, 'args': data[1]}
+                self.cur.send_sql('delete', send_data)
+                return self.cur.get_count()
+            except Exception as e:
+                # 发生错误
+                print(f'删除发生错误:{e}')
+                return False
 
-            # 删除数据库中指定内容
-            sql = f'delete from {table_name} where {data[0]} = %s;'
-            send_data = {'sql': sql, 'args': data[1]}
-            self.cur.send_sql('delete', send_data)
-            return self.cur.get_count()
-        except Exception as e:
-            # 发生错误
-            print(f'删除发生错误:{e}')
-            return False
+        if table_name == TABLE_NAME_FRIENDINFO:
+            try:
+                sql = f'delete from {table_name} where id = %s and friend = %s;'
+                send_data = {'sql': sql, 'args': data}
+                self.cur.send_sql('delete', send_data)
+                return self.cur.get_count()
+            except Exception as e:
+                # 发生错误
+                print(f'删除发生错误:{e}')
+                return False
 
     def edit(self, table_name, data):
         """
@@ -95,7 +129,11 @@ class ConnectSQL():
         :param data['待修改用户id'， '属性名'， '该属性修改后的内容']
         :return 修改后该用户的数据  未查找到相关用户则返回None, 查询失败则返回False
         """
-        for i in self.property_name:
+
+        if table_name == TABLE_NAME_FRIENDINFO:
+            print('friendinfo 不需要edit操作')
+            return False
+        for i in self.property_name_userinfo:
             if i == data[1]:
                 try:
                     if self.search(table_name, ['id', data[0]]) is None:
